requirements.txt
# Backend requirements for dbt Metadata Explorer
# Minimal dependencies required to run the application

# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data processing
PyYAML>=6.0
jinja2>=3.1.2

# API and Networking
httpx>=0.25.0
python-multipart>=0.0.6
requests>=2.28.2
starlette>=0.26.1

# Utilities
networkx>=3.1
typing-extensions>=4.5.0

# File handling and monitoring
watchdog>=3.0.0          # File system monitoring for auto-refresh

# Optional: Database support (uncomment if needed)
# sqlalchemy>=2.0.0
# pymysql>=1.0.0
# psycopg2-binary>=2.9.0

# Optional: DBT support (uncomment if needed)
# dbt-core>=1.5.0
# dbt-duckdb>=1.5.0
# dbt-sqlite>=1.5.1

# Optional: AI descriptions (uncomment if using AI features)
# google-generativeai>=0.3.1
# openai>=0.28.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Development
black==23.9.1
isort==5.12.0
flake8==6.1.0
mypy==1.5.1

# Command line tools
click>=8.1.3             # Command line interface creation

# Typing extensions
typing-extensions>=4.5.0 